chore: create solution
Created a basic working solution to use for the example, it contains a client and server web api
with docker-compose already configured so you can just clone and run. Server is on port 8000 and
client is on port 8001. Port 5001 is configured on the server but not used yet, this will be the
port we'll use for gRPC.